Relocate gateway source files and fix example
