+++ examples/basic/package.json
@@ -19,7 +19,7 @@
   },
   "dependencies": {
     "@graphql-tools/utils": "9.1.3",
-    "@moleculer-graphql/moleculer-graphql": "workspace:^",
+    "@moleculer-graphql/gateway": "workspace:^",
     "@moleculer-graphql/service": "workspace:^",
     "graphql": "16.6.0",
     "graphql-depth-limit": "1.1.0",
