+++ packages/gateway/src/GatewayStitcher.ts
