--- packages/gateway/src/classes/index.ts
@@ -1,2 +1,0 @@
-export { default as GatewayStitcher } from './GatewayStitcher';
-export { default as RequestHandler, type Request } from './RequestHandler';
