--- packages/gateway/src/mixins/index.ts
@@ -1,1 +1,0 @@
-export { default as gatewayMixin, type GatewayMixinOptions } from './gatewayMixin';
